requirements.txt
alembic==1.13.2
bcrypt==4.2.1
blinker==1.8.2
click==8.1.7
coverage==7.6.10
dnspython==2.7.0
email_validator==2.2.0
factory_boy==3.3.1
Faker==33.0.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.2.0
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.23.1
marshmallow-sqlalchemy==1.1.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
psycopg2-binary==2.9.9
pyclean==3.0.0
PyJWT==2.10.1
pytest==8.3.3
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
SQLAlchemy==2.0.32
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.0.3
WTForms==3.1.2
zipp==3.20.0
